refactor: 또 파일 세팅 적용
