--- packages/student/.vite/deps/@iconify_react_dist_iconify__js.js
@@ -1,40 +1,0 @@
-"use client";
-import {
-  Icon,
-  InlineIcon,
-  _api,
-  addAPIProvider,
-  addCollection,
-  addIcon,
-  calculateSize,
-  disableCache,
-  enableCache,
-  getIcon,
-  iconLoaded,
-  iconToSVG,
-  listIcons,
-  loadIcon,
-  loadIcons,
-  replaceIDs
-} from "./chunk-7FF6OZHY.js";
-import "./chunk-6ENKNRI7.js";
-export {
-  Icon,
-  InlineIcon,
-  _api,
-  addAPIProvider,
-  addCollection,
-  addIcon,
-  iconToSVG as buildIcon,
-  calculateSize,
-  disableCache,
-  enableCache,
-  getIcon,
-  iconLoaded as iconExists,
-  iconLoaded,
-  listIcons,
-  loadIcon,
-  loadIcons,
-  replaceIDs
-};
-//# sourceMappingURL=@iconify_react_dist_iconify__js.js.map
